Update App.tsx
In order to avoid nested urls, you have to remove the leading forward slash from the urls. This will connect the urls to the root url which is the homepage instead of combining different urls.

For eg:- If you have an edit url for editing notes. After clicking the edit button you'll be redirected to "localhost/edit-note" instead of "localhost/note-detail/edit-note".